Moves CoreDataFeedStore implementation of the FeedStore protocol to separate file
