+++ FeedModules/Core Layer/Tests/FeedCoreModuleTests/Feed Cache/Helpers/ImageDataStoreSpy.swift
@@ -8,22 +8,22 @@
 import Foundation
 import FeedCoreModule
 
-class ImageDataStoreSpy: ImageDataStore {
+class ImageDataStoreSpy: ImageStore {
     enum Message: Equatable {
         case insert(data: Data, for: URL)
         case retrieve(dataFor: URL)
     }
     
     private(set) var receivedMessages = [Message]()
-    private var retrievalCompletions = [(ImageDataStore.RetrievalResult) -> Void]()
-    private var insertionCompletions = [(ImageDataStore.InsertionResult) -> Void]()
+    private var retrievalCompletions = [(ImageStore.RetrievalResult) -> Void]()
+    private var insertionCompletions = [(ImageStore.InsertionResult) -> Void]()
     
-    func retrieve(dataForURL url: URL, completion: @escaping (ImageDataStore.RetrievalResult) -> Void) {
+    func retrieve(dataForURL url: URL, completion: @escaping (ImageStore.RetrievalResult) -> Void) {
         receivedMessages.append(.retrieve(dataFor: url))
         retrievalCompletions.append(completion)
     }
     
-    func insert(_ data: Data, for url: URL, completion: @escaping (ImageDataStore.InsertionResult) -> Void) {
+    func insert(_ data: Data, for url: URL, completion: @escaping (ImageStore.InsertionResult) -> Void) {
         receivedMessages.append(.insert(data: data, for: url))
         insertionCompletions.append(completion)
     }
